Move root to base folder
